+++ packages/jest-cli/src/get_changed_files_promise.js
@@ -21,6 +21,7 @@ export default (
       [],
     );
     return getChangedFilesForRoots(allRootsForAllProjects, {
+      changedSince: globalConfig.changedSince,
       lastCommit: globalConfig.lastCommit,
       withAncestor: globalConfig.changedFilesWithAncestor,
     });