+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## master
 
+### Features
+* `[jest-cli]` `--changedSince`: allow selectively running tests for code
+  changed since arbitrary revisions.
+  ([#5312](https://github.com/facebook/jest/pull/5312))
+
 ## jest 22.1.3
 
 ### Fixes